+++ src/com/jpaulmorrison/jdbcstuff/resources/layouts/Book.java
@@ -1,9 +1,11 @@
 package com.jpaulmorrison.jdbcstuff.resources.layouts;
 
+import java.math.BigDecimal;
+
 public class Book {
 	public String title;
 	public String author;
-	public double price;
+	public BigDecimal price;
 	public int    qty;
 }
